+++ crates/mimic/src/core/types/mod.rs
@@ -1,3 +1,4 @@
+mod account;
 mod blob;
 mod date;
 mod decimal;
@@ -13,6 +14,7 @@ mod timestamp;
 mod ulid;
 mod unit;
 
+pub use account::*;
 pub use blob::*;
 pub use date::*;
 pub use decimal::*;