+++ crates/mimic/src/design/base/types/ic/mod.rs
@@ -4,34 +4,6 @@ pub mod icrc3;
 use crate::design::prelude::*;
 
 ///
-/// Account
-///
-
-#[record(fields(
-    field(ident = "owner", value(item(prim = "Principal"))),
-    field(ident = "subaccount", value(opt, item(prim = "Subaccount")))
-))]
-pub struct Account {}
-
-impl Account {
-    pub fn new<P: Into<Principal>, S: Into<Subaccount>>(owner: P, subaccount: Option<S>) -> Self {
-        Self {
-            owner: owner.into(),
-            subaccount: subaccount.map(Into::into),
-        }
-    }
-}
-
-impl<P: Into<Principal>> From<P> for Account {
-    fn from(p: P) -> Self {
-        Self {
-            owner: p.into(),
-            ..Default::default()
-        }
-    }
-}
-
-///
 /// Memo
 ///